pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "document-checker-tool"
version = "0.1.0"
description = "A tool for checking and validating documents"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "documentcheckertool"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
gradio = "4.19.2"
python-docx = "1.1.2"
nltk = "3.9.1"
pandas = "2.2.1"
numpy = "1.24.3"
pydantic = "2.11.4"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-cov = "4.1.0"

[tool.ruff]
line-length = 120
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []
exclude = [
    ".git",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
    "venv",
]

[tool.ruff.isort]
known-first-party = ["documentcheckertool"]
known-third-party = ["gradio", "docx", "colorama", "nltk", "pydantic"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=documentcheckertool --cov-report=term-missing"

[tool.coverage.run]
source = ["documentcheckertool"]
omit = [
    "*/tests/*",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if False:",
    "if TYPE_CHECKING:",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.vulture]
exclude = ["tests/*", "setup.py"]
min-confidence = 100